+++ Kernel/include/BlockMemoryManager.h
@@ -10,9 +10,9 @@ createMemoryManagerImpl(void *const restrict memoryForMemoryManager,
 void initManagerImpl(MemoryManagerADT manager);
 void *allocMemoryImpl(MemoryManagerADT manager, size_t size);
 void freeMemoryImpl(MemoryManagerADT manager, void *ptr);
-void memStateImpl(MemoryManagerADT const restrict memoryManager, int * freeMemory,
-                  int * totalMemory, int * allocatedMemory);
+void memStateImpl(MemoryManagerADT const restrict memoryManager, size_t * freeMemory,
+                  size_t * totalMemory, size_t * allocatedMemory);
 
-void memStateRec(int * freeMemory, int * allocatedMemory, BlockADT currentBlock);
+void memStateRec(size_t * freeMemory, size_t * allocatedMemory, BlockADT currentBlock);
 
 #endif // BLOCK_MEMORY_MANAGER_H
