+++ Kernel/include/BuddyMemoryManager.h
@@ -14,6 +14,6 @@ createMemoryManagerImpl(void *const restrict memoryForMemoryManager,
 void initManagerImpl(MemoryManagerADT manager);
 void *allocMemoryImpl(MemoryManagerADT manager, size_t size);
 void freeMemoryImpl(MemoryManagerADT manager, void *ptr);
-void memStateImpl(MemoryManagerADT const restrict memoryManager, int * freeMemory, int * totalMemory, int * allocatedMemory);
+void memStateImpl(MemoryManagerADT const restrict memoryManager, size_t * freeMemory, size_t * totalMemory, size_t * allocatedMemory);
 
 #endif // BUDDY_MEMORY_MANAGER_H
